+++ backend/src/utils/chat/models.rs
@@ -1,7 +1,9 @@
 use serde::{Deserialize, Serialize};
 use time::OffsetDateTime;
+use typeshare::typeshare;
 use uuid::Uuid;
 
+#[typeshare]
 #[derive(Serialize, Deserialize, Debug)]
 pub struct AddresedMessage {
     pub content: String,
@@ -9,6 +11,7 @@ pub struct AddresedMessage {
     pub group_id: Uuid,
 }
 
+#[typeshare]
 #[derive(Serialize, Deserialize, Debug)]
 pub struct GroupUserMessageModel {
     pub nickname: String,
@@ -16,11 +19,12 @@ pub struct GroupUserMessageModel {
     pub sent_at: OffsetDateTime,
 }
 
+#[typeshare]
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct GroupUserMessage {
     pub nickname: String,
     pub content: String,
-    pub sat: i64,
+    pub sat: i32,
 }
 
 impl GroupUserMessage {
@@ -28,11 +32,12 @@ impl GroupUserMessage {
         Self {
             nickname,
             content,
-            sat: OffsetDateTime::now_utc().unix_timestamp(),
+            sat: OffsetDateTime::now_utc().unix_timestamp() as i32,
         }
     }
 }
 
+#[typeshare]
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct KickMessage {
     pub from: String,
