+++ backend/src/utils/groups/models.rs
@@ -1,17 +1,21 @@
 use serde::{Deserialize, Serialize};
+use typeshare::typeshare;
 use uuid::Uuid;
 
+#[typeshare]
 #[derive(Serialize, Deserialize)]
 pub struct NewGroup {
     pub name: String,
 }
 
+#[typeshare]
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Group {
     pub id: Uuid,
     pub name: String,
 }
 
+#[typeshare]
 #[derive(Serialize, Deserialize, Debug, Clone, Copy, PartialEq)]
 pub struct GroupUser {
     pub user_id: Uuid,
@@ -24,13 +28,15 @@ impl GroupUser {
     }
 }
 
+#[typeshare]
 #[derive(Deserialize)]
 pub struct NewGroupInvitation {
     pub group_id: Uuid,
 }
 
+#[typeshare]
 #[derive(Deserialize, Serialize, Debug, PartialEq)]
 pub struct GroupInfo {
     pub name: String,
-    pub members: i64,
+    pub members: i32,
 }
